+++ packages/core/src/index.ts
@@ -13,6 +13,7 @@ export type {
   SlotProps,
   VTableGuildOptions,
   VTableGuildContext,
+  VTableGuildThemeOverridesMap,
   ThemePresetName,
   VTableGuildLocale,
   VTableGuildTableLocale,